refactor: 머지 후 안쓰는 토큰 상태 삭제 및 타입 정리
